Fix AngelScript header compilation error
'asSMethodPtr::Convert' fails to compile due to an invalid return type.
However, an unspecialized version of the method is never meant to be
instantiated, in which case a default definition is not unnecessary. If
code ever attempts to instantiate a unspecialized version of the method,
compilation will fail due to the missing method definition.